finishing up

+++ app/Http/Controllers/ListingController.php
@@ -45,7 +45,7 @@ class ListingController extends Controller
         if($request->hasFile('logo')){
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         };
-
+        $formFields['user_id'] = auth()->id();
         Listing::create($formFields);
 
         return redirect('/')->with('message','Listing created successfully!');
@@ -59,6 +59,11 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing){
+
+        // Make sure the logged in user is the owner of the listing
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -81,7 +86,19 @@ class ListingController extends Controller
 
 
     public function destroy(Listing $listing) {
+
+        // Make sure the logged in user is the owner of the listing
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message','Listing deleted Successfully');
+    }
+
+    // Manage Listing function
+
+    public function manage(){
+        return view('listings.manage',['listings'=> auth()->user()->listings()->get()]);
     }
 }
